Add mix state_tests.run task
What changed?
============

Adds `mix state_tests.run` task to enable us to run a single state test
file quickly. The mix task also takes an optional `--fork` flag to only
run tests in that hardfork.

This task is the counterpart to the blockchain task `mix
common_tests.run` for state tests.

Since this mix task uses much of the same logic as the
`test/blockchain/state_test.exs` we extract that into `StateTestRunner`
module.
